+++ resources/views/indexes/categories.blade.php
@@ -17,48 +17,13 @@
       <b>{{ session('status') }}</b>
   </div>
   @endif
-  @if(!$is_regional)
-  <section class="jumbotron text-center">
-    <div class="container">
-      <h1>Répertoire de ressources locales<br>en contexte de distanciation sociale</h1>
-      <p class="lead text-muted">Une initiative citoyenne, en collaboration<br>avec plusieurs partenaires locaux</p>
-      <p>
-        <a href="{{ route('places.create-public') }}" class="btn btn-primary my-2">Inscrivez une entreprise</a>
-        <a href="{{ route('map.show') }}" class="btn btn-primary my-2"><i class="fas fa-map-marker-alt"></i> Voir la carte interactive</a>
-      </p>
-    </div>
-  </section>
-  @endif
 
   <div class="album py-5 bg-light">
     <div class="container">
 
-      @if($is_regional)
-      <h1 class="text-center mb-2">{{ $selectedRegion->name }}</h1>
-      @endif
-
-      @if(isset($category))
-      <h2 class="text-center mb-5">{{ $category->name }}</h2>
-      @endif
+      <h2 class="text-center mb-5">{{ $selectedRegion->name }} : {{ $category->name }}</h2>
 
       <div class="row" id="region-list">
-      @if($is_search)
-        <div class="col-md-12 text-center mb-5 h5">
-          <h3 class="mb-4">Filtrer par région</h3>
-          <a href="{{ route("recherche.index") }}{{ $q ? '?q='.$q : '' }}" class="badge badge-info">Tout le Québec <span class="badge badge-light">{{ App\Place::countByKeyword($q) }}</span></a>
-          @foreach(App\Region::all() as $region)
-          <a href="{{ route("recherche.index-region", ['region' => $region->slug]) }}{{ $q ? '?q='.$q : '' }}" class="badge badge-info">{{ $region->name }} <span class="badge badge-light">{{ $region->countPlacesByKeyword($q) }}</span></a>
-          @endforeach
-        </div>
-        @elseif(!$is_regional)
-        <div class="col-md-12 text-center mb-5 h5">
-          <h3 class="mb-4">Filtrer par région</h3>
-          <a href="{{ route("regions.index-provincial") }}" class="badge badge-info">Tout le Québec <span class="badge badge-light">{{ App\Place::where('is_approved', true)->count() }}</span></a>
-          @foreach(App\Region::all() as $region)
-          <a href="{{ route("regions.index-region", ['region' => $region->slug]) }}" class="badge badge-info">{{ $region->name }} <span class="badge badge-light">{{ $region->places()->where('is_approved', true)->count() }}</span></a>
-          @endforeach
-        </div>
-        @else
         <div class="col-md-12 text-center mb-5 h5">
           <a href="{{ route('regions.index-region', ['region' => $selectedRegion]) }}" class="badge badge-info">Toutes catégories <span class="badge badge-light">{{ App\Place::where('region_id', $selectedRegion->id)->where('is_approved', true)->count() }}</span></a>
             @foreach($categories as $category)
@@ -67,7 +32,6 @@
                 @endif
             @endforeach
         </div>
-        @endif
       </div>
 
       @if (session('status'))
@@ -78,10 +42,6 @@
 
       @if($places->isEmpty())
       <div class="alert alert-info">Toujours aucune entreprise enregistrée dans cette région! Vous en connaissez une? <b><a href="{{ route('places.create-public') }}">Inscrivez-là!</a></b></div>
-      @endif
-
-      @if(!$is_regional && !isset($category) && !$is_provincial && !$is_search)
-      <h3 class="mb-4 text-center">Quelques exemples</h3>
       @endif
 
       <div class="col-md-12" id="result-actions">