+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Changed icon for close and show place.
 - Moved about text to its own page.
 - Changed header bar, moved search to it.
+- Refactored indexes template.
 
 # 2.8.0
 - Fixed page title in social share meta.